% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickBLASTdeps.R
\name{isQuickBLASTLoaded}
\alias{isQuickBLASTLoaded}
\title{Stub function that always returns true. Only to test the connection of DLLs and C function calls}
\usage{
isQuickBLASTLoaded()
}
\value{
Always TRUE
}
\description{
Stub function that always returns true. Only to test the connection of DLLs and C function calls
}
\examples{
\dontrun{
QuickBLASTdeps::isQuickBLASTLoaded()
}

}
